Add option to assign custom values to cleared fields and properties
Now ClearOnReload attribute can be given a custom value to assign to field/property on reload instead of null. Plus, it can now be instructed to initialize the field/property with a new instance of its value type (helpful for static fields with collections, which have to be re-initialized on each reload to avoid NullReferenceExceptions.

Also added attribute constructors with doc comments to simplify their usage.